Completed a sample python language server to demo hover and code lens and updated the npm package list and pipenv package list. (#11)
