+++ Pipfile
@@ -8,8 +8,6 @@ mypy = "*"
 
 [packages]
 python-jsonrpc-server = "*"
-tornado = "*"
-websockets = "*"
 
 [requires]
 python_version = "3.7"
