+++ client/package.json
@@ -14,9 +14,7 @@
         "vscode": "^1.22.0"
     },
     "dependencies": {
-        "vscode-languageclient": "^6.1.3",
-        "@types/ws": "^7.2.5",
-        "ws": "^7.3.0"
+        "vscode-languageclient": "^6.1.3"
     },
     "devDependencies": {
         "@types/vscode": "^1.14.0",
