+++ package.json
@@ -75,8 +75,6 @@
 		"glob": "^7.1.6",
 		"mocha": "^7.1.2",
 		"typescript": "^3.8.3",
-		"vscode-test": "^1.3.0",
-		"@types/ws": "^7.2.5",
-		"ws": "^7.3.0"
+		"vscode-test": "^1.3.0"
 	}
 }
