--- python_server/socket_server.py
@@ -1,122 +1,0 @@
-import logging
-import os
-import socketserver
-import threading
-from functools import partial
-from pyls_jsonrpc.dispatchers import MethodDispatcher
-from pyls_jsonrpc.endpoint import Endpoint
-from pyls_jsonrpc.streams import JsonRpcStreamReader, JsonRpcStreamWriter
-
-
-log = logging.getLogger(__name__)
-
-
-MAX_WORKERS = 64
-
-
-class _StreamHandlerWrapper(socketserver.StreamRequestHandler, object):
-	"""A wrapper class that is used to construct a custom handler class."""
-
-	delegate = None
-
-	def setup(self):
-		super(_StreamHandlerWrapper, self).setup()
-		# pylint: disable=no-member
-		self.delegate = self.DELEGATE_CLASS(self.rfile, self.wfile)
-
-	def handle(self):
-		try:
-			self.delegate.start()
-		except OSError as e:
-			if os.name == 'nt':
-				# Catch and pass on ConnectionResetError when parent process
-				# dies
-				# pylint: disable=no-member, undefined-variable
-				if isinstance(e, WindowsError) and e.winerror == 10054:
-					pass
-
-		# pylint: disable=no-member
-		self.SHUTDOWN_CALL()
-
-
-def start_tcp_lang_server(bind_addr, port, check_parent_process, handler_class):
-	if not issubclass(handler_class, PythonLanguageServer):
-		raise ValueError('Handler class must be an instance of PythonLanguageServer')
-
-	def shutdown_server(check_parent_process, *args):
-		# pylint: disable=unused-argument
-		if check_parent_process:
-			log.debug('Shutting down server')
-			# Shutdown call must be done on a thread, to prevent deadlocks
-			stop_thread = threading.Thread(target=server.shutdown)
-			stop_thread.start()
-
-	# Construct a custom wrapper class around the user's handler_class
-	wrapper_class = type(
-		handler_class.__name__ + 'Handler',
-		(_StreamHandlerWrapper,),
-		{'DELEGATE_CLASS': partial(handler_class,
-								   check_parent_process=check_parent_process),
-		 'SHUTDOWN_CALL': partial(shutdown_server, check_parent_process)}
-	)
-
-	server = socketserver.TCPServer((bind_addr, port), wrapper_class, bind_and_activate=False)
-	server.allow_reuse_address = True
-
-	try:
-		server.server_bind()
-		server.server_activate()
-		log.info('Serving %s on (%s, %s)', handler_class.__name__, bind_addr, port)
-		server.serve_forever()
-	finally:
-		log.info('Shutting down')
-		server.server_close()
-
-
-class PythonLanguageServer(MethodDispatcher):
-	"""Implement a JSON RPC method dispatcher for the language server protocol."""
-
-	def __init__(self, rx, tx, check_parent_process=False):
-		# Endpoint is lazily set after construction
-		# self.endpoint = None
-		self._jsonrpc_stream_reader = JsonRpcStreamReader(rx)
-		self._jsonrpc_stream_writer = JsonRpcStreamWriter(tx)
-		self._check_parent_process = check_parent_process
-		self._endpoint = Endpoint(self, self._jsonrpc_stream_writer.write, max_workers=MAX_WORKERS)
-
-	def start(self):
-		"""Entry point for the server."""
-		self._jsonrpc_stream_reader.listen(self._endpoint.consume)
-
-	def m_initialize(self, rootUri=None, **kwargs):
-		log.info("Initialising custom Python MRO language server.")
-		return {"capabilities": {
-			"textDocumentSync": {
-				"openClose": True,
-				"change": 2,  # TextDocumentSyncKind.Incremental = 2
-			},
-			# "codeLensProvider": {
-			# 	"resolveProvider": True,
-			# },
-			"hoverProvider": True,
-		}}
-	
-	def m_hover(self, hoverParams=None, **_kwargs):
-		log.info("Providing Hover in {0} for line {1} char {2}".format(
-			hoverParams['textDocument']['uri'],
-			hoverParams['position']['line'],
-			hoverParams['position']['character'],
-		))
-		self._endpoint.notify('textDocument/hover', {
-			'contents': [
-				'Target class name',
-				'Parent class 1',
-				'Parent class 2',
-				'...',
-				'Object',
-			],
-		})
-
-
-if __name__ == "__main__":
-	start_tcp_lang_server('127.0.0.1', 3000, False, PythonLanguageServer)
